misc(client): added scenes folder and files
