+++ client/src/scenes/connection.cpp
@@ -1,0 +1,17 @@
+/**
+ * @file connection.cpp
+ * @brief Handles the rendering of the connection scene
+ * @date 09-01-2024
+ */
+
+#include "client.hpp"
+
+/**
+ * @brief Draws the connection screen
+ *
+ */
+void rtype::Client::drawConnection()
+{
+    _graphical->clear();
+    _graphical->display();
+}
