+++ client/src/scenes/gameOver.cpp
@@ -1,0 +1,15 @@
+/**
+ * @file gameOver.cpp
+ * @brief Handles the rendering of the game over scene
+ * @date 09-01-2024
+ */
+
+#include "client.hpp"
+
+/**
+ * @brief Draws the game over screen
+ *
+ */
+void rtype::Client::drawEnd()
+{
+}
